tests: Move crate following tests to `krate::following` module
