+++ src/tests/krate/mod.rs
@@ -13,6 +13,7 @@ use diesel::{dsl::*, prelude::*, update};
 
 mod dependencies;
 mod downloads;
+mod following;
 mod publish;
 mod summary;
 mod versions;
@@ -632,62 +633,6 @@ fn yanked_versions_are_not_considered_for_max_version() {
     let json = anon.search("q=foo");
     assert_eq!(json.meta.total, 1);
     assert_eq!(json.crates[0].max_version, "1.0.0");
-}
-
-#[test]
-fn diesel_not_found_results_in_404() {
-    let (_, _, user) = TestApp::init().with_user();
-
-    user.get("/api/v1/crates/foo_following/following")
-        .assert_not_found();
-}
-
-#[test]
-fn following() {
-    // TODO: Test anon requests as well?
-    let (app, _, user) = TestApp::init().with_user();
-
-    app.db(|conn| {
-        CrateBuilder::new("foo_following", user.as_model().id).expect_build(conn);
-    });
-
-    let is_following = || -> bool {
-        #[derive(Deserialize)]
-        struct F {
-            following: bool,
-        }
-
-        user.get::<F>("/api/v1/crates/foo_following/following")
-            .good()
-            .following
-    };
-
-    let follow = || {
-        assert!(
-            user.put::<OkBool>("/api/v1/crates/foo_following/follow", b"")
-                .good()
-                .ok
-        );
-    };
-
-    let unfollow = || {
-        assert!(
-            user.delete::<OkBool>("api/v1/crates/foo_following/follow")
-                .good()
-                .ok
-        );
-    };
-
-    assert!(!is_following());
-    follow();
-    follow();
-    assert!(is_following());
-    assert_eq!(user.search("following=1").crates.len(), 1);
-
-    unfollow();
-    unfollow();
-    assert!(!is_following());
-    assert_eq!(user.search("following=1").crates.len(), 0);
 }
 
 #[test]
